refactor: Remove redundant usage of useMemo and separate utility fn to utils.ts
